moved existing implementation to legacy provider
