+++ sdk/identity/identity/test/internal/node/workloadIdentityCredential.spec.ts
@@ -72,7 +72,6 @@ describe("WorkloadIdentityCredential", function () {
   it("authenticates with ManagedIdentity Credential", async function (this: Context) {
     process.env.AZURE_FEDERATED_TOKEN_FILE = tokenFilePath;
     const credential = new ManagedIdentityCredential("dummy-clientId");
-    assert.equal(credential["clientId"], "dummy-clientId");
     const token = await credential.getToken(scope);
     assert.ok(token?.token);
     assert.ok(token?.expiresOnTimestamp! > Date.now());
